vdr-1.4.5

+++ font.h
@@ -4,12 +4,13 @@
  * See the main source file 'vdr.c' for copyright information and
  * how to reach the author.
  *
- * $Id: font.h 1.12 2006/02/05 13:46:36 kls Exp $
+ * $Id: font.h 1.14 2006/12/03 16:42:10 kls Exp $
  */
 
 #ifndef __FONT_H
 #define __FONT_H
 
+#include <stdint.h>
 #include <stdlib.h>
 
 enum eDvbFont {
@@ -32,7 +33,7 @@ enum eDvbCode {
 class cFont {
 public:
   enum { NUMCHARS = 256 };
-  typedef unsigned long tPixelData;
+  typedef uint32_t tPixelData;
   struct tCharData {
     tPixelData width, height;
     tPixelData lines[1];
